Refactor Readium CSS
